:construction: updated StartGame and removed bis version
