+++ apps/web/src/components/functional/StartGame.tsx
@@ -1,3 +1,5 @@
+import type { CreateGameDto } from '@shifumi/dtos'
+import { createGame } from '../../services/api'
 import BasicButton from '../ui/BasicButton'
 
 interface StartGameProps {
@@ -6,10 +8,24 @@ interface StartGameProps {
   buttonTitle: string
 }
 function StartGame({ onClick, isStarted, buttonTitle }: StartGameProps) {
+  const handleStartGame = async () => {
+    try {
+      const newGame: CreateGameDto = {
+        playerOneId: 1,
+        playerTwoId: 2,
+      }
+      const createdGame = await createGame(newGame)
+      console.warn('Jeu créé :', createdGame)
+      onClick()
+    }
+    catch (error) {
+      console.error('Erreur lors de la création du jeu :', error)
+    }
+  }
   return (
     <div>
       {!isStarted && (
-        <BasicButton onClick={onClick} buttonTitle={buttonTitle} />
+        <BasicButton onClick={handleStartGame} buttonTitle={buttonTitle} />
       )}
     </div>
   )
